+++ Tracker/IExpenses.cs
@@ -5,7 +5,7 @@ namespace Tracker
 {
 	public interface IExpenses
 	{
-		void Add (int amount, Record.Types type, Record.Categories category, string description);
+		void Add (decimal amount, Record.Types type, Record.Categories category, string description);
 		ObservableCollection<Record> Records { get; }
 	}
 }
